frontend/requirements.txt
streamlit==1.39.0
python-socketio==5.11.4
websocket-client==1.8.0